web/requirements.txt
Flask
yt-dlp
Flask-Cors
python-dotenv